refactor: split monolithic Python script into a few smaller modules
